feat(coach): Allow coaches to manage multiple teams
Coaches can now be assigned to multiple teams, enabling them to oversee and manage players and training sessions across different age groups or squads. This is achieved by changing the `teamId` field to `teamIds` (an array) in the `User` type and updating the UI logic in `PlayerManager` and `TrainingPlanner` to accommodate this new structure. Mock users and relevant components have also been updated to reflect this change.